debut ajout POST Client

+++ src/main/java/com/accenture/applicationlocationvehicule/repository/AdressDao.java
@@ -3,5 +3,5 @@ package com.accenture.applicationlocationvehicule.repository;
 import com.accenture.applicationlocationvehicule.repository.entity.Adress;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdressDao extends JpaRepository<Adress , Integer> {
+public interface AdressDao extends JpaRepository<Adress , Long> {
 }
